Move node types into 'types' module
